refactor: 락 컴포넌트 네이밍 변경 및 패키지 위치 이동
