+++ src/main/kotlin/kr/hhplus/be/server/api/reservation/application/ReservationFacade.kt
@@ -12,9 +12,9 @@ import kr.hhplus.be.server.domain.reservation.model.CreateReservation
 import kr.hhplus.be.server.domain.reservation.model.Reservation
 import kr.hhplus.be.server.domain.support.error.CoreException
 import kr.hhplus.be.server.domain.support.error.ErrorType
-import kr.hhplus.be.server.domain.support.lock.LockResource
-import kr.hhplus.be.server.domain.support.lock.LockStrategy
-import kr.hhplus.be.server.domain.support.lock.aop.DistributedLock
+import kr.hhplus.be.server.domain.support.component.lock.LockResource
+import kr.hhplus.be.server.domain.support.component.lock.LockStrategy
+import kr.hhplus.be.server.domain.support.component.lock.aop.DistributedLock
 import kr.hhplus.be.server.listener.event.ReservationCompleteEvent
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Component
