+++ src/main/kotlin/kr/hhplus/be/server/domain/balance/BalanceService.kt
@@ -2,9 +2,9 @@ package kr.hhplus.be.server.domain.balance
 
 import kr.hhplus.be.server.domain.balance.model.Balance
 import kr.hhplus.be.server.domain.balance.model.ModifyBalance
-import kr.hhplus.be.server.domain.support.lock.LockResource
-import kr.hhplus.be.server.domain.support.lock.LockStrategy
-import kr.hhplus.be.server.domain.support.lock.aop.DistributedLock
+import kr.hhplus.be.server.domain.support.component.lock.LockResource
+import kr.hhplus.be.server.domain.support.component.lock.LockStrategy
+import kr.hhplus.be.server.domain.support.component.lock.aop.DistributedLock
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.math.BigDecimal
