+++ src/main/kotlin/kr/hhplus/be/server/domain/support/component/lock/LockHandler.kt
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.domain.support.component.lock
+
+interface LockHandler {
+    fun unlock()
+}
