+++ src/main/kotlin/kr/hhplus/be/server/domain/support/component/lock/LockResource.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.support.lock
+package kr.hhplus.be.server.domain.support.component.lock
 
 enum class LockResource(
     private val description: String,
