+++ src/main/kotlin/kr/hhplus/be/server/domain/support/component/lock/LockStrategy.kt
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.domain.support.lock
+package kr.hhplus.be.server.domain.support.component.lock
 
 enum class LockStrategy(
     val clientName: String,
