--- src/main/kotlin/kr/hhplus/be/server/domain/support/lock/LockHandler.kt
@@ -1,5 +1,0 @@
-package kr.hhplus.be.server.domain.support.lock
-
-interface LockHandler {
-    fun unlock()
-}
